+++ ChangeLog.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `gradient_normalization` argument that can be `element` or `path`
+- `path_normalization` can be a number between 0 and 1
+
 ## [0.6.3] - 2022-06-29
 ### Added
 - Docker image